Add machinery to check for PR-blocking-todos during CI (#11)
This gives one the opportunity to "block" pull requests by editing a markdown file.

If there are any TODO items found in the `.pr-blocking-todos.md` file, a CI check visibly fails. Of course one could ignore this and merge the pull request anyway, but at least this serves as a visible warning.